fixed the dummy image issue

+++ src/services/user/user-services.ts
@@ -486,7 +486,7 @@ export const profileServices = {
   updateUser: async (payload: any) => {
     if (payload.image) {
       const data = await UserModel.findById(payload.id).lean();
-      if (data && data.image) {
+      if (data && data.image && data.image !== "admin/images/image+(2).png") {
         await deleteFileFromS3(data?.image as string);
       }
     }
